Remix_monolit/requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
matplotlib==3.7.2
Pillow==10.0.0
numpy==1.24.3
pdfrw
reportlab
boto3
natsort
google-cloud-storage==2.10.0
google-auth

# Mannequin Segmenter dependencies
transformers>=4.21.0
torch>=1.12.0
torchvision>=0.13.0
pillow>=9.0.0
opencv-python-headless>=4.5.0
matplotlib>=3.5.0
boto3>=1.26.0

# Category Predictor dependencies
scikit-learn>=1.1.0
joblib>=1.1.0

# Attribute Predictor dependencies
openai>=1.0.0
python-dotenv>=0.19.0